add custom exceptions

+++ src/main/java/ma/uca/ensas/ecommercebackendspring/service/CityService.java
@@ -2,6 +2,7 @@ package ma.uca.ensas.ecommercebackendspring.service;
 
 import lombok.RequiredArgsConstructor;
 import ma.uca.ensas.ecommercebackendspring.entities.City;
+import ma.uca.ensas.ecommercebackendspring.exceptions.ApiRequestException;
 import ma.uca.ensas.ecommercebackendspring.repositories.CityRepository;
 import org.springframework.stereotype.Service;
 
@@ -18,10 +19,16 @@ public class CityService {
     }
 
     public City saveCity(City city){
+        if (cityRepository.findByName(city.getName()).isPresent()){
+            throw new ApiRequestException("City with name : " + city.getName() + " already exists !");
+        }
         return cityRepository.save(city);
     }
 
     public void deleteCityById(Long id){
+        if (cityRepository.findById(id).isEmpty()){
+            throw new ApiRequestException("City with id : " + id + " doesn't exists");
+        }
         cityRepository.deleteById(id);
     }
 }
